Move agent init to init.rs
